+++ service/processor/__mocks__/IncomingEmailProcessor.ts
@@ -1,0 +1,8 @@
+export class IncomingEmailProcessor {
+  // Static properties to access the mocks from your tests
+  static mockProcessEmail = jest.fn();
+
+  async processEmail(...args) {
+    return IncomingEmailProcessor.mockProcessEmail(...args);
+  }
+}
